+++ src/app/Fake.Sql.SqlPackage/Sql.SqlPackage.fs
@@ -47,17 +47,22 @@ module SqlPackage =
         Profile : string }
 
     let internal validPaths =
-        let getSqlVersion (path:string) = path.Split '\\' |> Array.item 3 |> int
-        let getVsVersion (path: string) = (Path.GetDirectoryName path |> DirectoryInfo).Name |> int
-        let sql = !!(Environment.ProgramFilesX86 </> @"Microsoft SQL Server\**\DAC\bin\SqlPackage.exe") |> Seq.map(fun path -> path, getSqlVersion path)
-        let vsOld32Bit = !!(Environment.ProgramFilesX86 </> @"Microsoft Visual Studio*\Common7\IDE\Extensions\Microsoft\SQLDB\DAC\*\SqlPackage.exe") |> Seq.map(fun path -> path, getVsVersion path)
-        let vsModern32Bit = !!(Environment.ProgramFilesX86 </> @"Microsoft Visual Studio\**\Common7\IDE\Extensions\Microsoft\SQLDB\DAC\*\SqlPackage.exe") |> Seq.map(fun path -> path, getVsVersion path)
-        let vs64bit =
-            !!(Environment.ProgramFiles </> @"Microsoft Visual Studio\**\Common7\IDE\Extensions\Microsoft\SQLDB\DAC\SqlPackage.exe")
-            |> Seq.map(fun path -> path, Reflection.Assembly.LoadFile(path).GetName().Version.Major)
+        let paths = [
+            let macOrLinux = Set [ PlatformID.MacOSX; PlatformID.Unix ]
+            if macOrLinux.Contains Environment.OSVersion.Platform then
+                !!"/usr/local/bin/sqlpackage"
+                |> Seq.map (fun path -> path, 15)
+            else
+                let getSqlVersion (path:string) = path.Split '\\' |> Array.item 3 |> int
+                let getVsVersion (path: string) = (Path.GetDirectoryName path |> DirectoryInfo).Name |> int
+                !!(Environment.ProgramFilesX86 </> @"Microsoft SQL Server\**\DAC\bin\SqlPackage.exe") |> Seq.map(fun path -> path, getSqlVersion path)
+                !!(Environment.ProgramFilesX86 </> @"Microsoft Visual Studio*\Common7\IDE\Extensions\Microsoft\SQLDB\DAC\*\SqlPackage.exe") |> Seq.map(fun path -> path, getVsVersion path)
+                !!(Environment.ProgramFilesX86 </> @"Microsoft Visual Studio\**\Common7\IDE\Extensions\Microsoft\SQLDB\DAC\*\SqlPackage.exe") |> Seq.map(fun path -> path, getVsVersion path)
+                !!(Environment.ProgramFiles </> @"Microsoft Visual Studio\**\Common7\IDE\Extensions\Microsoft\SQLDB\DAC\SqlPackage.exe") |> Seq.map(fun path -> path, Reflection.Assembly.LoadFile(path).GetName().Version.Major)
+        ]
 
-        [ sql; vsOld32Bit; vsModern32Bit; vs64bit ]
-        |> Seq.collect id
+        paths
+        |> Seq.concat
         |> Seq.sortByDescending snd
         |> Seq.map fst
         |> Seq.cache